requirements.txt
Flask==2.2.3
Flask_Mail==0.9.1
mysql_connector_repackaged==0.3.1
Pillow==9.4.0
stability_sdk==0.4.0
